Migrate Shipping resources in table-rate-shipping to split-class layout
  ShippingZoneResource           310 -> 75 lines  (+ShippingZoneForm 195, +ShippingZoneTable 50)
  ShippingMethodResource         278 -> 90 lines  (+ShippingMethodForm 175, +ShippingMethodTable 50)
  ShippingExclusionListResource  122 -> 70 lines  (+ShippingExclusionListForm 40, +ShippingExclusionListTable 45)

Cross-page consumers updated in lockstep:
  - ListShippingZones, ListShippingMethod, ListShippingExclusionLists →
    {Resource}Form::get*Component
  - ManageShippingExclusions → ShippingExclusionListTable::getColumns()

ShippingZone's private syncPostcodes helper moves to ShippingZoneForm.
TaxZone-style: helpers tightly coupled to form save closures stay with
the form class.

Closes the Shipping family. Every Lunar resource is now on the
split-class layout — the wrapper traits on BaseResource have no
remaining callers.
