+++ packages/table-rate-shipping/src/Filament/Resources/ShippingExclusionListResource.php
@@ -2,20 +2,15 @@
 
 namespace Lunar\Shipping\Filament\Resources;
 
-use Filament\Actions\BulkActionGroup;
-use Filament\Actions\DeleteBulkAction;
-use Filament\Actions\EditAction;
-use Filament\Forms;
 use Filament\Pages\Enums\SubNavigationPosition;
-use Filament\Schemas\Components\Component;
-use Filament\Schemas\Components\Section;
 use Filament\Schemas\Schema;
 use Filament\Support\Facades\FilamentIcon;
-use Filament\Tables\Columns\TextColumn;
 use Filament\Tables\Table;
 use Lunar\Admin\Support\Resources\BaseResource;
 use Lunar\Shipping\Filament\Resources\ShippingExclusionListResource\Pages;
 use Lunar\Shipping\Filament\Resources\ShippingExclusionListResource\RelationManagers\ShippingExclusionRelationManager;
+use Lunar\Shipping\Filament\Resources\ShippingExclusionListResource\Schemas\ShippingExclusionListForm;
+use Lunar\Shipping\Filament\Resources\ShippingExclusionListResource\Tables\ShippingExclusionListTable;
 use Lunar\Shipping\Models\Contracts\ShippingExclusionList;
 
 class ShippingExclusionListResource extends BaseResource
@@ -48,13 +43,14 @@ class ShippingExclusionListResource extends BaseResource
         return __('lunarpanel.shipping::plugin.navigation.group');
     }
 
-    public static function getDefaultForm(Schema $schema): Schema
+    public static function form(Schema $schema): Schema
     {
-        return $schema->components([
-            Section::make()->schema(
-                static::getMainFormComponents(),
-            ),
-        ]);
+        return ShippingExclusionListForm::configure($schema);
+    }
+
+    public static function table(Table $table): Table
+    {
+        return ShippingExclusionListTable::configure($table);
     }
 
     protected static function getDefaultRelations(): array
@@ -64,55 +60,7 @@ class ShippingExclusionListResource extends BaseResource
         ];
     }
 
-    protected static function getMainFormComponents(): array
-    {
-        return [
-            static::getNameFormComponent(),
-        ];
-    }
-
-    public static function getNameFormComponent(): Component
-    {
-        return Forms\Components\TextInput::make('name')
-            ->label(__('lunarpanel.shipping::shippingexclusionlist.form.name.label'))
-            ->required()
-            ->maxLength(255)
-            ->autofocus();
-    }
-
-    public static function getDefaultTable(Table $table): Table
-    {
-        return $table
-            ->columns(static::getTableColumns())
-            ->filters([
-                //
-            ])
-            ->actions([
-                EditAction::make(),
-            ])
-            ->bulkActions([
-                BulkActionGroup::make([
-                    DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getTableColumns(): array
-    {
-        return [
-            TextColumn::make('name')
-                ->label(
-                    __('lunarpanel.shipping::shippingexclusionlist.table.name.label')
-                ),
-            TextColumn::make('exclusions_count')
-                ->label(
-                    __('lunarpanel.shipping::shippingexclusionlist.table.exclusions_count.label')
-                )
-                ->counts('exclusions'),
-        ];
-    }
-
-    public static function getDefaultPages(): array
+    protected static function getDefaultPages(): array
     {
         return [
             'index' => Pages\ListShippingExclusionLists::route('/'),
