+++ packages/table-rate-shipping/src/Filament/Resources/ShippingMethodResource/Pages/ListShippingMethod.php
@@ -7,6 +7,7 @@ use Filament\Schemas\Components\Group;
 use Lunar\Admin\Support\Pages\BaseListRecords;
 use Lunar\Core\Models\CustomerGroup;
 use Lunar\Shipping\Filament\Resources\ShippingMethodResource;
+use Lunar\Shipping\Filament\Resources\ShippingMethodResource\Schemas\ShippingMethodForm;
 use Lunar\Shipping\Models\ShippingMethod;
 
 class ListShippingMethod extends BaseListRecords
@@ -17,13 +18,13 @@ class ListShippingMethod extends BaseListRecords
     {
         return [
             Actions\CreateAction::make()->form([
-                ShippingMethodResource::getNameFormComponent(),
+                ShippingMethodForm::getNameComponent(),
                 Group::make([
-                    ShippingMethodResource::getCodeFormComponent(),
-                    ShippingMethodResource::getDriverFormComponent(),
-                    ShippingMethodResource::getChargeByFormComponent(),
+                    ShippingMethodForm::getCodeComponent(),
+                    ShippingMethodForm::getDriverComponent(),
+                    ShippingMethodForm::getChargeByComponent(),
                 ])->columns(2),
-                ShippingMethodResource::getDescriptionFormComponent(),
+                ShippingMethodForm::getDescriptionComponent(),
             ])->after(function (ShippingMethod $shippingMethod) {
                 $customerGroups = CustomerGroup::pluck('id')->mapWithKeys(
                     fn ($id) => [$id => ['visible' => true, 'enabled' => true, 'starts_at' => now()]]
